roll over iso time period tests to utils from workflow
